added array to separate file to import
