+++ locales/en.js
@@ -19,6 +19,7 @@ export default {
   joinRoomNamePlaceholder: 'Your name',
   joinRoomButton: 'Join Room',
   required: ' * ',
+  selectCards: 'Select your cards',
 
   // Room
   users: 'Users',
