+++ locales/fr.js
@@ -21,6 +21,7 @@ export default {
   joinRoomNamePlaceholder: 'Votre nom',
   joinRoomButton: 'Rejoignez la salle',
   required: ' * ',
+  selectCards: 'Sélectionnez vos cartes',
 
   // Room
   users: '(FR)Users',
