+++ utils/cards.js
@@ -1,0 +1,24 @@
+export const cards = [
+  { id: 'card-0', src: '/Card_0.svg', alt: 'Card image', value: 0 },
+  { id: 'card-0.5', src: '/Card_0.5.svg', alt: 'Card image', value: 1000 },
+  { id: 'card-1', src: '/Card_1.svg', value: 1 },
+  { id: 'card-2', src: '/Card_2.svg', alt: 'Card image', value: 2 },
+  { id: 'card-3', src: '/Card_3.svg', alt: 'Card image', value: 3 },
+  { id: 'card-4', src: '/Card_5.svg', alt: 'Card image', value: 5 },
+  { id: 'card-5', src: '/Card_8.svg', alt: 'Card image', value: 8 },
+  { id: 'card-6', src: '/Card_13.svg', alt: 'Card image', value: 13 },
+  { id: 'card-7', src: '/Card_20.svg', alt: 'Card image', value: 20 },
+  { id: 'card-8', src: '/Card_40.svg', alt: 'Card image', value: 40 },
+  { id: 'card-9', src: '/Card_100.svg', alt: 'Card image', value: 100 },
+  {
+    id: 'card-10',
+    src: '/Card_questionMark.svg',
+    alt: 'Card image',
+    value: 1001,
+  },
+  { id: 'card-11', src: '/Card_infinity.svg', alt: 'Card image', value: 1002 },
+]
+
+export default {
+  cards,
+}
